pyproject.toml
[tool.poetry]
name = "fluent-mind-bot"
version = "0.1.0"
description = ""
authors = ["Andrey Ivanov <ivnv.xd@gmail.com>"]
readme = "README.md"
packages = [{include = "config"}]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.21.1"
python-telegram-bot = "^20.1"
openai = "^0.27.0"
Django = "^4.1.7"
gunicorn = "^20.1.0"
django-bootstrap4 = "^22.3"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pytest = "^7.2.1"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
bot = "bot.app:main"
